+++ src/com/tagbio/umap/metric/CanberraMetric.java
@@ -17,10 +17,10 @@ public class CanberraMetric extends Metric {
   }
 
   @Override
-  public double distance(final float[] x, final float[] y) {
-    double result = 0.0;
+  public float distance(final float[] x, final float[] y) {
+    float result = 0;
     for (int i = 0; i < x.length; ++i) {
-      final double denominator = Math.abs(x[i]) + Math.abs(y[i]);
+      final float denominator = Math.abs(x[i]) + Math.abs(y[i]);
       if (denominator > 0) {
         result += Math.abs(x[i] - y[i]) / denominator;
       }
