+++ src/com/tagbio/umap/metric/ChebyshevMetric.java
@@ -17,9 +17,9 @@ public class ChebyshevMetric extends Metric {
   }
 
   @Override
-  public double distance(final float[] x, final float[] y) {
+  public float distance(final float[] x, final float[] y) {
     // D(x, y) = \max_i |x_i - y_i|
-    double result = 0.0;
+    float result = 0;
     for (int i = 0; i < x.length; ++i) {
       result = Math.max(result, Math.abs(x[i] - y[i]));
     }
