+++ src/com/tagbio/umap/metric/MatchingMetric.java
@@ -17,7 +17,7 @@ public class MatchingMetric extends Metric {
   }
 
   @Override
-  public double distance(final float[] x, final float[] y) {
+  public float distance(final float[] x, final float[] y) {
     int numNotEqual = 0;
     for (int i = 0; i < x.length; ++i) {
       final boolean xTrue = x[i] != 0;
